% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scgeyser.r
\name{scGeyser}
\alias{scGeyser}
\title{scGeyser}
\usage{
scGeyser(mgif_dir = NULL)
}
\arguments{
\item{mgif_dir}{Optional directory containing mgif yaml files. If given, then the UI will display 
the yaml files in a new UI section on the load screen.}
}
\description{
Run the shiny app for single-cell data visualization.
}
