% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scgeyser_module_data_loader.R
\name{dataLoaderServer}
\alias{dataLoaderServer}
\title{Server for the Data Loader Module}
\usage{
dataLoaderServer(id, mgif_dir = NULL)
}
\arguments{
\item{id}{Namespace ID.}

\item{mgif_dir}{Optional local directory for MGIF files.}
}
\value{
A reactive list containing `$config`, `$obs_data`, `$gene_table`, and a data access function `$get_gene_data`.
}
\description{
Handles the logic for loading data from different sources like MGIF, RDS, and HDF5.
}
