% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scgeyser_module_gene_table.R
\name{geneTableServer}
\alias{geneTableServer}
\title{Server for the Gene Table Module}
\usage{
geneTableServer(id, gene_data, config)
}
\arguments{
\item{id}{Namespace ID.}

\item{gene_data}{A reactive data.table of gene information.}

\item{config}{The configuration information specifying file inputs and slot information}
}
\value{
A reactive containing the name of the gene selected from the table.
}
\description{
A reusable server component for the gene table that returns a list of the selected gene (both the file name and the displayed name)
}
